requirements.txt
email-validator>=2.2.0
faiss-cpu>=1.10.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.32.3
pydantic>=2.9.2
pathway>=0.20.1
flask-socketio>=5.5.1
numpy>=2.2.3
langchain>=0.3.20
langchain-core>=0.3.44
